Reorganise loader directory into src directory
