+++ src/package.json
@@ -1,0 +1,12 @@
+{
+  "dependencies": {
+    "comlink": "^4.3.1"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.25.0",
+    "esbuild": "^0.14.39",
+    "eslint": "^8.15.0",
+    "eslint-config-google": "^0.14.0",
+    "typescript": "^4.6.4"
+  }
+}
