+++ src/templates/console.html
@@ -1,0 +1,8 @@
+<!DOCTYPE html>
+<html>
+<head>
+</head>
+<body>
+</body>
+<script type="module" src="console.js"></script>
+</html>
