+++ src/templates/package.json
